create the edition slice
